+++ packages/ui/src/form/primitives/searchable-chip-select/index.ts
@@ -3,6 +3,7 @@ import { ChipWithRemove } from '../combobox/chip-with-remove';
 import { Group } from '../combobox/group';
 import { GroupLabel } from '../combobox/group-label';
 import { Collection } from '../combobox/collection';
+import { useFilteredItems } from '../combobox/use-filtered-items';
 import { SearchableChipSelect as _SearchableChipSelect } from './searchable-chip-select';
 
 Item.displayName = 'SearchableChipSelect.Item';
@@ -15,6 +16,8 @@ GroupLabel.displayName = 'SearchableChipSelect.GroupLabel';
  * support for a creatable footer action.
  *
  * Prefer `SearchableChipSelectControl` when using with a standard label and description.
+ *
+ * Announces a visually hidden result count as the list filters.
  */
 export const SearchableChipSelect = Object.assign( _SearchableChipSelect, {
 	Item,
@@ -22,4 +25,5 @@ export const SearchableChipSelect = Object.assign( _SearchableChipSelect, {
 	Group,
 	GroupLabel,
 	Collection,
+	useFilteredItems,
 } );